+++ client/src/surfaces/chat/ChatPanel.tsx
@@ -1,4 +1,5 @@
 import { ChatComposer } from "./ChatComposer.js";
+import { ChatDebugModeToggle } from "./ChatDebugModeToggle.js";
 import { ChatHeader } from "./ChatHeader.js";
 import { MessageList } from "./MessageList.js";
 import { RunStatusBanner } from "./RunStatusBanner.js";
@@ -10,6 +11,7 @@ export function ChatPanel() {
       <RunStatusBanner />
       <MessageList />
       <ChatComposer />
+      <ChatDebugModeToggle />
     </section>
   );
 }